Hadoop Reindex Task from Coordinator
Fixes https://github.com/druid-io/druid/issues/1517

Deprecates druid.coordinator.merge.on. Instead, user will use druid.coordinator.merge.strategy for segment merging.

In this PR,
 ```DruidCoordinatorHadoopSegmentMerger``` will find small segments in the segment timeline, and submit HadoopIndexTask to reindex those imbalanced segments.

For example,
Here is a timeline of segments for a specific dataSource (Note: The segment granularity for each segment could be arbitrary, there might also be a gap between two segments)
```
SegID:    1   2   3   4   5   6   7   8   9  10  11  12   13
Shard0: |_O_|_O_|_O_|_O_|_S_|_S_|_S_|_O_|_S_|_O_|_S_|_O_|_S_|__|__|__|__|__|__|__|
Shard1:         |_S_|                   |_S_|   |_S_|_S_|
Shard2:         |_S_|                   |_S_|   |_S_|

S: the size of shard is below configured optimal size
O: the size of shard >= configured optimal size
```
As ```DruidCoordinatorHadoopSegmentMerger``` runs periodically, eventually there will be no shards in the timeline that are tagged with S.

Here is the algorithm ```DruidCoordinatorHadoopSegmentMerger``` implements for finding imbalanced segments and submits the reindex task,

```
unbalancedIntervals = []
currTotalSize = 0
intervalToReindex = null
shouldBeMerged = false  // whether we should reindex segments within intervalToReindex

// the direction of scanning will alternate if we don't want to merge segments
// that have gaps between them
for each segment

  expand intervalToReindex with the interval of current segment

  for each shard of current segment
    if shard.getSize() < optimal size
      shouldBeMerged = true // found a small segment
    currTotalSize += shard.getSize()

  // this is an estimate that whether we can get a good-size segment by
  // reindexing the segments within intervalToReindex
  if currTotalSize >= optimal size
    if shouldBeMerged
      unbalancedIntervals.add(intervalToReindex)
    currTotalSize = 0
    intervalToReindex = null
    shouldBeMerged = false
  else {
    // do nothing. Since we don't have enough total size to make a segment
    // with size >= optimal size, we'll seek to merge it with the next segment in the timeline
  }

if unbalancedIntervals is not empty:
  submit HadoopIndexTask with unbalancedIntervals

if we want to keep the segment gap during merge
  invert the scan direction so that in next Coordinator's run, we will scan the segments in the inverted direction.
Note: The reason we need to invert the scan direction when keepGap is enabled
is because it's possible to have a small segment that is at the boundary of gap left over
 and never gets merged.

```

Basically, the algorithm iterates each segment in the timeline, and for each shard in that segment, we check if its size is below optimal size, if yes, we should definitely try to merge that shard with others. However, it's possible that the total size of shards in that segment is not big enough to make an optimal size segment, so we should seek to merge the current segment with the next one in the timeline.
Note: If we want to keep the gap between segments that don't abut, it's possible to have small segments left over, however, we can invert the scan direction in the next Coordinator's run to eliminate those small segments.